+++ packages/celer-web-app/src/data/bundler/index.ts
@@ -1,4 +1,3 @@
 export * from "./RouteScriptBundler";
 export * from "./switch";
 export * from "./type";
-export * from "./version";