+++ packages/celer-web-app/src/data/docs/functions.ts
@@ -1,15 +1,13 @@
-import { TARGET_VERSION } from "data/bundler";
 
 export const exampleRouteScriptFunctions = {
-	compilerVersion: TARGET_VERSION,
-	Project: {
-		Name: "Function Examples",
-		Authors: ["iTNTPiston"],
-		Url: "https://celer.itntpiston.app/#/docs/functions",
-		Version: "1.0.0",
-		Description: "Examples of functions"
+	_project: {
+		name: "Function Examples",
+		authors: ["iTNTPiston"],
+		url: "https://celer.itntpiston.app/#/docs/functions",
+		version: "1.0.0",
+		description: "Examples of functions"
 	},
-	Route: [
+	_route: [
 		{
 			"Example Functions":[
 				".link(https://youtu.be/ZYAnULQZyiw?t=73)",